requirements.txt
asgiref==3.8.1
distro==1.9.0
Django==4.2
gunicorn==21.2.0
JPype1==1.5.0
numpy==1.24.2
packaging==24.0
pandas==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tabula-py==2.9.0
tzdata==2024.1
whitenoise==6.6.0
XlsxWriter==3.2.0
